requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageshow.py: 2
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 27
# C:\Users\phill\Documents\coed_thing\expression_system_AI\videoshow.py: 2
Pillow == 8.1.0

# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageee.py: 7
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 13,14,21
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 7,8,15
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 14,15,22
matplotlib == 3.3.3

# C:\Users\phill\Documents\coed_thing\expression_system_AI\create_dataser.py: 8
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageee.py: 6
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageshow.py: 9
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 6
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 5
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 7
# C:\Users\phill\Documents\coed_thing\expression_system_AI\videoshow.py: 9
numpy == 1.19.4

# C:\Users\phill\Documents\coed_thing\expression_system_AI\haarcascade_detect.py: 1
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageee.py: 5
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageshow.py: 3
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 11
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 4
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 12
# C:\Users\phill\Documents\coed_thing\expression_system_AI\videoshow.py: 3
opencv_python == 4.4.0.46

# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 12
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 6
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 13
pandas == 1.2.0

# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 24
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 18
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 25
sklearn == 0.0

# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 28
tensorboardX == 2.2

# C:\Users\phill\Documents\coed_thing\expression_system_AI\CNNnet.py: 1,2,3,4
# C:\Users\phill\Documents\coed_thing\expression_system_AI\create_dataser.py: 7
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageshow.py: 4,6,11
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 5,10,18,19,22,23,25
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 1,2,12,13,16,17,19
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 6,11,19,20,23,24,26
# C:\Users\phill\Documents\coed_thing\expression_system_AI\videoshow.py: 4,6,11
torch == 1.7.1

# C:\Users\phill\Documents\coed_thing\expression_system_AI\create_dataser.py: 6
# C:\Users\phill\Documents\coed_thing\expression_system_AI\imageshow.py: 5
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 16
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 10
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 17
# C:\Users\phill\Documents\coed_thing\expression_system_AI\videoshow.py: 5
torchvision == 0.8.2

# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 17
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 11
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 18
tqdm == 4.55.0

# C:\Users\phill\Documents\coed_thing\expression_system_AI\fuck.py: 1
# C:\Users\phill\Documents\coed_thing\expression_system_AI\newtrain.py: 15
# C:\Users\phill\Documents\coed_thing\expression_system_AI\test.py: 9
# C:\Users\phill\Documents\coed_thing\expression_system_AI\train.py: 16
yaspin == 1.5.0
